Implemented integration with Amazon RDS (PostgreSQL)

+++ src/main/java/aws/imgupload/imgupload/service/MetadataService.java
@@ -6,5 +6,5 @@ public interface MetadataService {
     void startRegister(ImageMetadata metadata);
     void commitRegister(ImageMetadata metadata);
     void rollbackRegister(ImageMetadata metadata);
-    int fetchImageCount();
+    long fetchImageCount();
 }
